Remove TODO comment and include Terraform scripts
Removed the TODO comment regarding the contact form from the index.html file as its issue was already addressed previously and it was no longer required.

Meanwhile, The Terraform scripts were added to manage the AWS infrastructure setup in AWS in a more automated and scalable way. It includes S3 bucket for hosting, Route53 for DNS, ACM for SSL certificate, and CloudFront for CDN.

This allows us to manage the infrastructure as code, ensures better tracking of changes and offers a more repeatable process when setting up new environments. Add yml conf for github actions